refactor: add each seed type to its respective file.
